requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
google-generativeai==0.8.3
google-cloud-aiplatform==1.66.0
requests==2.31.0


